requirements.txt
pyodbc>=4.0.35
pandas>=1.3.0
chardet>=5.0.0
python-dotenv>=0.19.0
packaging>=21.0 